+++ packages/extension-polkagate/src/fullscreen/stake/new-solo/Info.tsx
@@ -2,14 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { SoloStakingInfo } from '../../../hooks/useSoloStakingInfo';
-import type { PopupCloser, Stats } from '../util/utils';
+import type { PopupCloser } from '../util/utils';
 
 import { Container } from '@mui/material';
-import { Calendar, Discover, MagicStar } from 'iconsax-react';
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { GradientButton } from '../../../components';
-import { useChainInfo, useTranslation } from '../../../hooks';
+import { useChainInfo, useStakingInfoSolo, useTranslation } from '../../../hooks';
 import { DraggableModal } from '../../components/DraggableModal';
 import { InfoBox } from '../partials/InfoBox';
 
@@ -21,16 +20,9 @@ interface Props {
 
 export default function Info({ genesisHash, onClose, stakingInfo }: Props): React.ReactElement {
   const { t } = useTranslation();
-  const { decimal, token } = useChainInfo(genesisHash, true);
+  const { decimal } = useChainInfo(genesisHash, true);
 
-  const stakingStats: Stats[] = useMemo(() => ([
-    { InfoIcon: Discover, label: t('Max Validators you can select'), value: stakingInfo.stakingConsts?.maxNominations },
-    { label: t('Min {{token}} to be staker', { replace: { token: token ?? '' } }), value: stakingInfo.stakingConsts?.minNominatorBond, withLogo: true },
-    { label: t('Min {{token}} to receive rewards', { replace: { token: token ?? '' } }), value: 250, withLogo: true },
-    { InfoIcon: MagicStar, label: t('Max nominators of a validator, who may receive rewards'), value: stakingInfo.stakingConsts?.maxNominatorRewardedPerValidator },
-    { InfoIcon: Calendar, label: t('Days it takes to receive your funds back after unstaking'), value: stakingInfo.stakingConsts?.unbondingDuration },
-    { label: t('Min {{token}} that must remain in your account (ED)', { replace: { token: token ?? '' } }), value: 1, withLogo: true }
-  ]), [stakingInfo.stakingConsts?.maxNominations, stakingInfo.stakingConsts?.maxNominatorRewardedPerValidator, stakingInfo.stakingConsts?.minNominatorBond, stakingInfo.stakingConsts?.unbondingDuration, t, token]);
+  const stakingStats = useStakingInfoSolo(stakingInfo, genesisHash);
 
   return (
     <DraggableModal